fix pr checks
