+++ .npmrc
@@ -1,2 +1,3 @@
 save-exact=true
 engine-strict=true
+legacy-peer-deps=true
