+++ @types/app/index.d.ts
@@ -1,20 +1,2 @@
 import './global.d';
 
-declare global {
-    namespace Express {
-        interface Request {
-            /** Logger instance associated with current request */
-            logger: () => void;
-        }
-
-        interface Response {
-
-            /**
-             * Renders bundle to html, then sends it
-             * or performs redirect if necessary
-             */
-            // tslint:disable-next-line:no-any
-            renderBundle(bundleName: string, data?: any): void;
-        }
-    }
-}
