+++ client/actions.ts
@@ -1,4 +1,4 @@
-import store from "./utils/infrastructure/store";
+import { store } from './utils/infrastructure/store';
 
 export const pure = {};
 export const bound = store.bindActions(pure);
