+++ client/utils/infrastructure/reducers/entityLoader/types.ts
@@ -1,12 +1,10 @@
-import {Empty} from 'utils';
-
 import {CommonStore} from '../../store';
 import {BaseActions} from '../flow';
 
 export interface PaginationOptions {
     start: number; // Попросить бэк поменять на skip или offset
     limit?: number; // На бэке пока игнорируется
-    continue?: Empty<true>;
+    continue?: Nullable<true>;
 }
 
 export interface EntityLoaderConfig<D> {
