+++ package.json
@@ -35,6 +35,7 @@
     "@babel/preset-react": "7.18.6",
     "@babel/preset-typescript": "7.18.6",
     "@types/jest": "29.0.0",
+    "@types/node": "18.11.5",
     "@types/pg": "8.6.5",
     "@types/pkg-dir": "2.0.1",
     "@types/react": "18.0.14",
@@ -92,7 +93,7 @@
     "connected-react-router": "6.9.3",
     "dotenv": "16.0.1",
     "entity-api": "0.1.1",
-    "express": "4.17.3",
+    "express": "4.16.4",
     "formik": "2.2.9",
     "history": "5.3.0",
     "immer": "9.0.16",
