+++ server/app.ts
@@ -1,7 +1,7 @@
 import {config} from 'dotenv';
 import express, {Express} from 'express';
 
-import router from "../client/reducers/router";
+import { routerReducer } from "../client/reducers/router";
 import {notFound, queryParser} from "./controllers";
 
 
@@ -12,5 +12,5 @@ server
     .disable('x-powered-by')
     .enable('trust proxy')
     .set('query parser', queryParser)
-    .use(router)
+    .use(routerReducer)
     .use(notFound);
