+++ server/controllers/app.ts
@@ -1,13 +1,17 @@
 import {Request, Response} from 'express';
 
-import { cfg } from "../../lib/cfg";
-
-
-const RTC_ENV = process.env.RTC_ENV || cfg.__DEV__;
-
 export function renderApp(req: Request, res: Response) {
     const resHeaders = res.getHeaders();
+    // req.tld!
+
     const {
         ip,
-    } = req;
+        nonce,
+    } = (req as any);
+
+    res.renderBundle('desktop', {
+        ip,
+        nonce,
+        resHeaders,
+    });
 }
